// src/modules/quotes/enums/quote.enums.ts
export enum QuoteRequestStatus {
  ABIERTA = 'abierta',
  EN_COTIZACION = 'en_cotizacion',
  CERRADA = 'cerrada',
  CANCELADA = 'cancelada',
}

export enum OpportunityStatus {
  NUEVA = 'nueva',
  EN_ANALISIS = 'en_analisis',
  COTIZADA = 'cotizada',
  DESCARTADA = 'descartada',
}

export enum ProviderQuoteStatus {
  BORRADOR = 'borrador',
  ENVIADA = 'enviada',
  ACEPTADA = 'aceptada',
  RECHAZADA = 'rechazada',
  EXPIRADA = 'expirada',
}